Updated camera; added ball force
Updated the camera to now follow the ball and sit closer to the ball. Ball movement now applys a force to the ball. Issues with plane tilt can cause the ball to clip into the plane and fall into the abbyss. Still working out how to solve this, may only require a thicker plane but also might need updated tilting code. Tested with GameCube Controller and works correctly